add features
- Match typst files in subdirectories
- Add a way to get the typst files from a zip file
    - It extracts them into a temporary directory
    - Mostly to make it very easy to use the tool with a downloaded
archive from typst.app
- Add a cli interface
    - Option to ignore certain decks
    - Option to decide where to find the typst files
- Add a check for duplicate ids option: to do so, I had to separate the
parsing of the typst files from the creation of the anki cards
- Add progress bar
    - Big difficulty and design choice: I wanted to still be able to
interrupt the process while progress bars were displaying, but ctrl+c
can't be caught in the middle of a progress bar, because of the
rendering of the progress bar. **I haven't found a better solution, but
my current implementation isn't really clean. Improvements are
welcome.**
    - Wanted to keep the project zero-dependency, so I had to implement
my own progress bar; Maybe we could eventually allow a dependency for a
better progress bar implementation
- Add a dry-run mode, mostly for testing

+++ typ2anki/generator.py
@@ -1,6 +1,8 @@
 import os
 from pathlib import Path
 
+from typ2anki.config import config
+
 def ensure_ankiconf_file(directory):
     ankiconf_path = Path(directory) / "ankiconf.typ"
     if not ankiconf_path.exists():
@@ -10,13 +12,21 @@ def ensure_ankiconf_file(directory):
 ) = {
   doc
 }
-"""
+"""     
+        if config().dry_run:
+            print(f"Creating ankiconf file at {ankiconf_path}")
+            return
+        
         with open(ankiconf_path, "w") as file:
             file.write(default_content)
 
 def generate_card_file(card, card_id, output_path):
-    temp_file = "temporal.typ"
+    temp_file = Path(output_path) / "temporal.typ"
     output_file = Path(output_path) / f"{card_id}-{{p}}.png"
+
+    if config().dry_run:
+        print(f"Generating card file for card {card_id} at {output_file}")
+        return
 
     card_type = "custom-card" if "custom-card" in card else "card"
 
@@ -48,11 +58,10 @@ def generate_card_file(card, card_id, output_path):
 }}
 {card}
 """
-
+    
     try:
         with open(temp_file, "w") as file:
             file.write(template)
-
         os.system(f"typst c {temp_file} {output_file}")
     finally:
         if os.path.exists(temp_file):
